[REFACTOR] move updateReviewReply and deleteReviewReply to review.ts
